// ui/src/lib/types/api.types.ts
export interface components {
    schemas: {
// ...
        ProductConstraintsDto: {
            minAmount?: number;
            maxAmount?: number;
            minTerm?: number;
            maxTerm?: number;
            minRate?: number;
            maxRate?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}